hiremind-backend/AI_Model/requirements.txt
openai==1.109.1
pypdf2==3.0.1
python-docx==1.1.2
requests==2.32.3
pandas==2.1.4
python-dotenv==1.0.0
matplotlib==3.8.2

langchain==0.2.16
langchain-openai==0.1.25
streamlit==1.33.0